move into global utils package
